updates retated to class 108 https://www.youtube.com/watch?v=hkiTMaeICmw customized validation formarray and checkbox now it is completly functional and we transfered the validation method to a new class that we export as static so all modules can use the same method
